Move GitHub test reporter to validation pipeline
