+++ modules/validation-pipeline/github-test-reporter/src/main/kotlin/org/cafejojo/schaapi/validationpipeline/githubtestreporter/FuelConfiguration.kt
@@ -1,0 +1,12 @@
+package org.cafejojo.schaapi.validationpipeline.githubtestreporter
+
+import com.github.kittinunf.fuel.core.Client
+import com.github.kittinunf.fuel.core.FuelManager
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+internal open class FuelConfiguration(httpClient: Client) {
+    init {
+        FuelManager.instance.client = httpClient
+    }
+}
